requirements.txt
discord.py
motor
python-dotenv
aiohttp